backend/requirements.txt
annotated-types==0.6.0
Authlib==1.3.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
Flask-Cors==4.0.0
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gspread==6.1.0
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauth2client==4.1.3
oauthlib==3.2.2
pkg_resources==0.0.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
StrEnum==0.4.15
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
